refactor: LottoNumberStrategy를 FunctionalInterface로 선언하여 InputLottoNumberStrategy를 람다식으로 표현
