+++ src/main/java/lotto/domain/LottoMarket.java
@@ -1,6 +1,7 @@
 package lotto.domain;
 
 import lotto.domain.strategy.AutoLottoNumberStrategy;
+import lotto.domain.strategy.LottoNumberStrategy;
 
 import java.util.List;
 
@@ -8,8 +9,10 @@ public class LottoMarket {
 	private final Lottos lottos = new Lottos();
 
 	public LottoMarket(int numberOfLotto) {
+		LottoNumberStrategy strategy = new AutoLottoNumberStrategy();
+
 		for(int i = 0; i < numberOfLotto; i++) {
-			lottos.add(new Lotto(new AutoLottoNumberStrategy()));
+			lottos.add(new Lotto(strategy));
 		}
 	}
