+++ src/main/java/lotto/domain/strategy/LottoNumberStrategy.java
@@ -2,6 +2,7 @@ package lotto.domain.strategy;
 
 import java.util.List;
 
+@FunctionalInterface
 public interface LottoNumberStrategy {
 	int LOTTO_NUMBER_COUNT = 6;
 	int LOTTO_NUMBER_MIN = 1;
