+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,7 +1,6 @@
-package lotto;
+package lotto.domain;
 
 import lotto.domain.strategy.AutoLottoNumberStrategy;
-import lotto.domain.Lotto;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
