+++ src/test/java/lotto/domain/WinningNumbersTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.Test;
 
 import java.util.Set;
